view user profile

+++ src/main/java/com/takeandtrade/capstone/models/User.java
@@ -33,7 +33,7 @@ public class User {
     private String password;
 
     @Column (nullable = false)
-    private int phoneNumber;
+    private long phoneNumber;
 
     @Column (nullable = false, length = 255)
     private String city;
@@ -101,11 +101,11 @@ public class User {
         this.password = password;
     }
 
-    public int getPhoneNumber() {
+    public long getPhoneNumber() {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
 
@@ -146,7 +146,7 @@ public class User {
 
     public User(){}
 
-    public User(long id, String firstName, String lastName, String username, String dateOfBirth, String email, String password, int phoneNumber, String city, String state, int zipCode) {
+    public User(long id, String firstName, String lastName, String username, String dateOfBirth, String email, String password, long phoneNumber, String city, String state, int zipCode) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
